requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-crispy-forms==1.9.1
django-filter==2.2.0
django-login-required-middleware==0.5.0
django-timezone-field==4.0
kombu==4.6.8
mysqlclient==1.4.6
Pillow==7.1.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
